More database stuff
I did the approval and rejection of papers that have been submitted and
for now just to show functionality there is a link on the Homepage of
that (SEStatus) that takes you to a page where you can remove papers
that have already been accepted. At this point it bypasses the
moderation process but in the future I expect there to be a moderation
process between the choosing of the URL and the accept/reject screen.